+++ docs/montecarlointegration.tex
@@ -17,7 +17,7 @@
 \newcommand{\zo}[0]{{z_1}}
 \newcommand{\zt}[0]{{z_2}}
 
-\title{Convolutions using Monte-Carlo methods}
+\title{Convolution using Monte-Carlo methods}
 \author{Niccolò Laurenti}
 
 \date{\today}
