+++ inc/adani/CoefficientFunction.h
@@ -17,54 +17,10 @@
 #ifndef CoeffFunc
 #define CoeffFunc
 
-#include "adani/SplittingFunction.h"
+#include "adani/Value.h"
 
 #include <vector>
 #include <iostream>
-
-using std::vector;
-using std::ostream;
-
-class Value {
-    public:
-        Value(const double& central, const double& higher, const double& lower) ;
-        Value(const double& central);
-        Value(const double& higher, const double& lower) ;
-        Value(const Value& value) ;
-
-        double GetCentral() const {return central_;};
-        double GetHigher() const {return higher_;};
-        double GetLower() const {return lower_;};
-
-        vector<double> ToVect() const ;
-
-        Value operator+(const Value& rhs) const;
-        // Value operator-(const Value& rhs) const;
-
-        Value operator+(const double& rhs) const;
-        friend Value operator+(const double& lhs, const Value& rhs);
-
-        Value operator-(const double& rhs) const;
-        // friend Value operator-(const double& lhs, const Value& rhs);
-
-        Value operator*(const double& rhs) const;
-        friend Value operator*(const double& lhs, const Value& rhs);
-
-        Value operator/(const double& rhs) const;
-        friend Value operator/(const double& lhs, const Value& rhs);
-
-        const Value& operator=(const Value& rhs);
-
-        const Value& operator*=(const double& rhs);
-        const Value& operator/=(const double& rhs);
-
-        friend ostream& operator<<(ostream& os, const Value& rhs);
-
-    private:
-        double central_;
-        double higher_;
-        double lower_;
-};
 
 class CoefficientFunction {
