+++ inc/adani/MatchingCondition.h
@@ -1,10 +1,10 @@
 /*
  * =====================================================================================
  *
- *       Filename:  MatchingConditions.h
+ *       Filename:  MatchingCondition.h
  *
  *    Description:  Header file for the
- * MatchingConditions.cc file.
+ * MatchingCondition.cc file.
  *
  *         Author:  Vamo!!
  *
@@ -19,20 +19,7 @@
 #include "CoefficientFunction.h"
 
 //==========================================================================================//
-//  Matching conditions O(as)
-//------------------------------------------------------------------------------------------//
-
-double K_Qg1(double x, double m2mu2);
-double K_gg1_local(double m2mu2);
-
-//==========================================================================================//
-//  Matching conditions O(as^2)
-//------------------------------------------------------------------------------------------//
-
-double K_Qg2(double x, double m2mu2);
-
-//==========================================================================================//
-//  Matching conditions O(as^3)
+//  class MatchingCondition
 //------------------------------------------------------------------------------------------//
 
 class MatchingCondition {
@@ -49,6 +36,23 @@ class MatchingCondition {
         
         bool exact_;
         bool revised_approx_;
+
+        //==========================================================================================//
+        //  Matching conditions O(as)
+        //------------------------------------------------------------------------------------------//
+
+        // double K_Qg1(double x, double m2mu2) const;
+        // double K_gg1_local(double m2mu2) const ;
+
+        //==========================================================================================//
+        //  Matching conditions O(as^2)
+        //------------------------------------------------------------------------------------------//
+
+        // double K_Qg2(double x, double m2mu2) const ;
+
+        //==========================================================================================//
+        //  Matching conditions O(as^3)
+        //------------------------------------------------------------------------------------------//
 
         double a_Qg_30(double x, int v) const ;
         double a_Qq_PS_30(double x, int v) const ;
