+++ inc/adani/adani.h
@@ -1,19 +1,20 @@
 #ifndef adani_h
 #define adani_h
 
-#include "adani/ApproximateCoefficientFunctions.h"
+#include "adani/ApproximateCoefficientFunction.h"
+#include "adani/AsymptoticCoefficientFunction.h"
 #include "adani/CoefficientFunction.h"
-#include "adani/AsymptoticCoefficientFunctions.h"
 #include "adani/Constants.h"
-#include "adani/Convolutions.h"
-#include "adani/ExactCoefficientFunctions.h"
-#include "adani/HighEnergyCoefficientFunctions.h"
-#include "adani/HighScaleCoefficientFunctions.h"
+#include "adani/Convolution.h"
+#include "adani/ExactCoefficientFunction.h"
+#include "adani/HighEnergyCoefficientFunction.h"
+#include "adani/HighScaleCoefficientFunction.h"
 #include "adani/HighScaleSplitLogs.h"
-#include "adani/MasslessCoefficientFunctions.h"
-#include "adani/MatchingConditions.h"
+#include "adani/MasslessCoefficientFunction.h"
+#include "adani/MatchingCondition.h"
 #include "adani/SpecialFunctions.h"
 #include "adani/SplittingFunction.h"
 #include "adani/ThresholdCoefficientFunction.h"
+#include "adani/Value.h"
 
 #endif
