+++ src/AsymptoticCoefficientFunction.cc
@@ -1,6 +1,6 @@
-#include "adani/AsymptoticCoefficientFunctions.h"
-#include "adani/HighEnergyCoefficientFunctions.h"
-#include "adani/HighScaleCoefficientFunctions.h"
+#include "adani/AsymptoticCoefficientFunction.h"
+#include "adani/HighEnergyCoefficientFunction.h"
+#include "adani/HighScaleCoefficientFunction.h"
 #include <cmath>
 
 AsymptoticCoefficientFunction::AsymptoticCoefficientFunction(const int& order, const char& kind, const char& channel, const bool& NLL, const bool& exact_highscale, const bool& revised_approx_highscale) : AbstractHighEnergyCoefficientFunction(order, kind, channel, NLL) {
