+++ src/CoefficientFunction.cc
@@ -3,135 +3,6 @@
 using std::cout ;
 using std::endl ;
 
-Value::Value(const double& central, const double& higher, const double& lower) {
-    central_ = central;
-
-    if (higher < central) {
-        cout << "Error: class Value initialized with higher < central!" << endl;
-        cout << "Got: central=" << central << ", higher=" << higher << endl;
-        exit(-1);
-    }
-    higher_ = higher;
-
-    if (lower > central) {
-        cout << "Error: class Value initialized with lower > central!" << endl;
-        cout << "Got: central=" << central << ", lower=" << lower << endl;
-        exit(-1);
-    }
-    lower_ = lower;
-}
-
-Value::Value(const double& central) {
-    central_ = central;
-    higher_ = central;
-    lower_ = central;
-}
-
-Value::Value(const double& higher, const double& lower) {
-
-    if (higher < lower) {
-        cout << "Error: class Value initialized with higher < lower!" << endl;
-        cout << "Got: higher=" << higher << ", lower=" << lower << endl;
-        exit(-1);
-    }
-    higher_ = higher;
-    lower_ = lower;
-    central_ = (higher + lower) / 2;
-}
-
-Value::Value(const Value& value) {
-    central_ = value.central_;
-    higher_ = value.higher_;
-    lower_ = value.lower_;
-}
-
-vector<double> Value::ToVect() const {
-    return {central_, higher_, lower_};
-}
-
-Value Value::operator+(const Value& rhs) const {
-    return Value(central_ + rhs.central_, higher_ + rhs.higher_, lower_ + rhs.lower_);
-}
-
-// Value Value::operator-(const Value& rhs) const {
-    
-// }
-
-Value Value::operator+(const double& rhs) const {
-    return Value(rhs + central_, rhs + higher_, rhs + lower_);
-}
-
-Value operator+(const double& lhs, const Value& rhs){
-    return Value(lhs + rhs.central_, lhs + rhs.higher_, lhs + rhs.lower_);
-}
-
-Value Value::operator-(const double& rhs) const {
-    return Value(central_ - rhs, higher_ - rhs, lower_ - rhs);
-}
-
-// Value operator-(const double& lhs, const Value& rhs){
-//     return Value(lhs - rhs.central_, lhs - rhs.higher_, lhs - rhs.lower_);
-// }
-
-Value Value::operator*(const double& rhs) const {
-    if (rhs > 0) return Value(rhs * central_, rhs * higher_, rhs * lower_);
-    else return Value(rhs * central_, rhs * lower_, rhs * higher_);
-}
-
-Value operator*(const double& lhs, const Value& rhs){
-    return Value(rhs.central_, rhs.higher_, rhs.lower_) * lhs;
-}
-
-Value Value::operator/(const double& rhs) const {
-    if (rhs > 0) return Value(central_ / rhs, higher_ / rhs, lower_ / rhs);
-    else return Value(central_ / rhs, lower_ / rhs, higher_ / rhs);
-}
-
-Value operator/(const double& lhs, const Value& rhs){
-    return Value(rhs.central_, rhs.higher_, rhs.lower_) / lhs ;
-}
-
-const Value& Value::operator=(const Value& rhs) {
-    central_ = rhs.central_;
-    higher_ = rhs.higher_;
-    lower_ = rhs.lower_;
-    
-    return *this;
-}
-
-const Value& Value::operator*=(const double& rhs) {
-    if (rhs > 0) {
-        central_ *= rhs;
-        higher_ *= rhs;
-        lower_ *= rhs;
-    } else {
-        central_ *= rhs;
-        higher_ = lower_ * rhs;
-        lower_ = higher_ * rhs;  
-    }
-
-    return *this;
-}
-
-const Value& Value::operator/=(const double& rhs) {
-    if (rhs > 0) {
-        central_ /= rhs;
-        higher_ /= rhs;
-        lower_ /= rhs;
-    } else {
-        central_ /= rhs;
-        higher_ = lower_ / rhs;
-        lower_ = higher_ / rhs;  
-    }
-
-    return *this;
-}
-
-ostream& operator<<(ostream& os, const Value& rhs){
-    os << "(" << rhs.central_ << ", " << rhs.higher_ << ", " << rhs.lower_ << ")" ;
-    return os;
-}
-
 CoefficientFunction::CoefficientFunction(const int& order, const char& kind, const char& channel) {
 
     SetOrder(order);
