+++ src/HighScaleSplitLogs.cc
@@ -1,6 +1,6 @@
 #include "adani/HighScaleSplitLogs.h"
 #include "adani/Constants.h"
-#include "adani/Convolutions.h"
+#include "adani/Convolution.h"
 #include "adani/SpecialFunctions.h"
 
 #include <cmath>
