+++ src/MasslessCoefficientFunction.cc
@@ -1,4 +1,4 @@
-#include "adani/MasslessCoefficientFunctions.h"
+#include "adani/MasslessCoefficientFunction.h"
 #include "adani/Constants.h"
 #include "adani/SpecialFunctions.h"
